Superset frame bug fixed

+++ app/src/main/java/com/marcomadalin/olympus/data/database/converters/ListSetConverters.kt
@@ -8,13 +8,13 @@ import java.lang.reflect.Type
 
 object ListSetConverters {
     @TypeConverter
-    fun fromString(value: String): List<Set<Int>> {
-        val listType: Type = object : TypeToken<List<Set<Int>>>() {}.type
+    fun fromString(value: String): List<Set<Long>> {
+        val listType: Type = object : TypeToken<List<Set<Long>>>() {}.type
         return Gson().fromJson(value, listType)
     }
 
     @TypeConverter
-    fun fromListSet(list: List<Set<Int>>): String {
+    fun fromListSet(list: List<Set<Long>>): String {
         val gson = Gson()
         return gson.toJson(list)
     }
